fix(float): Validate window before closing
The window that is going to be closed (via `win:close()` call)
probably have already been closed or no longer available, in which case
"Invalid window id" errors will occur when closing the window.

To avoid such an error, we can if the window is valid right before
closing the window.