feat(files): custom find function
File searching is a tiny part of parsing time so there's no real benefit
to using external commands. Also found to be about the same or slightly
faster, guessing that's because most of the work is async and so the
same string manipulation is what is taking most of the time.

See #59