fix(watch): use old lsp supports method for v0.10.1 (#456)
* fix(watch): fixed neotest.watch starting problem

On the version v0.10.1, supports_method is still a table, rather than
a function. Code was trying to call a regular table field as a function,
resulting in following error:

attempt to call field 'supports_method' (a table value)

Now the type of the client.supports_method is checked beforehand,
instead of accessing the method right away.

* fix(watch): edge case in case neither client is found

* fix(watch): instead of type, checking the nvim version instead

Basically 1:1 code @rcasia proposed.