fix(dap): provide a safe context for dap initialization
If an adapter has started an async task before yielding back to neotest
(e.g. reading or writing a file), dap could be required in a fast event
context, resulting in a failed initialization. Ensure async tasks are
synchronized with the main loop to provide dap a safe initialization
context.